% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover.R
\name{get_discovery_feed}
\alias{get_discovery_feed}
\title{Get discovery items.}
\usage{
get_discovery_feed(
  page = 0,
  feed = "top",
  week = "this.week",
  genre = NULL,
  subgenre = NULL,
  format = NULL,
  location = NULL
)
}
\arguments{
\item{page}{Page to get. 0 is first page.}

\item{feed}{Feed category. One of \code{"top"} (best-selling), \code{"new"}
(new arrivals) or \code{"rec"} (artist-recommended).}

\item{week}{Week to get feed for. Must be one of \code{"today"},
\code{"this.week"} or a numeric week number.}

\item{genre}{Filter on genre e.g. \code{"jazz"}.
See Bandcamp front page for options.}

\item{subgenre}{Filter on subgenre e.g. \code{"bebop"}.
See Bandcamp front page for options.}

\item{format}{Filter on release format.
Must be one of \code{"digital"}, \code{"vinyl"}, \code{"cd"} or \code{"cassette"}.}

\item{location}{Filter on location e.g. \code{"paris"}.
See Bandcamp front page for options.}
}
\description{
Get discovery items.
}
