% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fan.R
\name{get_fan_collection}
\alias{get_fan_collection}
\title{Get data for all items in fan collection}
\usage{
get_fan_collection(username)
}
\arguments{
\item{username}{Username of the fan account}
}
\description{
Get data for all items in fan collection
}
