% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fan.R
\name{get_fan_followers}
\alias{get_fan_followers}
\title{Get data for all fan followers}
\usage{
get_fan_followers(username)
}
\arguments{
\item{username}{Username of the fan account}
}
\description{
Get data for all fan followers
}
