% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fan.R
\name{get_fan_wishlist}
\alias{get_fan_wishlist}
\title{Get data for all items in fan wishlist}
\usage{
get_fan_wishlist(username)
}
\arguments{
\item{username}{Username of the fan account}
}
\description{
Get data for all items in fan wishlist
}
